move object stats util func to disttae.util.go
