+++ pkg/vm/engine/disttae/util.go
@@ -18,6 +18,9 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"github.com/matrixorigin/matrixone/pkg/catalog"
+	"github.com/matrixorigin/matrixone/pkg/objectio"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/options"
 	"math"
 	"strings"
 
@@ -1297,4 +1300,119 @@ func ListTnService(appendFn func(service *metadata.TNService)) {
 		}
 		return true
 	})
+}
+
+// util function for object stats
+
+// UnfoldBlkInfoFromObjStats constructs a block info list from the given object stats.
+// this unfolds all block info at one operation, if an object contains a great many of blocks,
+// this operation is memory sensitive, we recommend another way, StatsBlkIter or ForEach.
+func UnfoldBlkInfoFromObjStats(stats *objectio.ObjectStats) (blks []catalog.BlockInfo) {
+	if stats.IsZero() {
+		return blks
+	}
+
+	name := stats.ObjectName()
+	blkCnt := uint16(stats.BlkCnt())
+	rowTotalCnt := stats.Rows()
+	accumulate := uint32(0)
+
+	for idx := uint16(0); idx < blkCnt; idx++ {
+		blkRows := uint32(options.DefaultBlockMaxRows)
+		if idx == blkCnt-1 {
+			blkRows = rowTotalCnt - accumulate
+		}
+		accumulate += blkRows
+		loc := objectio.BuildLocation(name, stats.Extent(), blkRows, idx)
+		blks = append(blks, catalog.BlockInfo{
+			BlockID:   *objectio.BuildObjectBlockid(name, idx),
+			MetaLoc:   catalog.ObjectLocation(loc),
+			SegmentID: name.SegmentId(),
+		})
+	}
+
+	return blks
+}
+
+// ForeachBlkInObjStatsList receives an object stats list,
+// and visits each blk of these object stats by OnBlock,
+// until the onBlock returns false or all blks have been enumerated.
+// when onBlock returns a false,
+// the nextStats argument decides whether continue onBlock on the next stats or exit foreach completely.
+func ForeachBlkInObjStatsList(
+	nextStats bool,
+	onBlock func(blk *catalog.BlockInfo) bool, statsList ...objectio.ObjectStats) {
+	stop := false
+	statsCnt := len(statsList)
+
+	for idx := 0; idx < statsCnt && !stop; idx++ {
+		iter := NewStatsBlkIter(&statsList[idx])
+		for iter.Next() {
+			blk := iter.Entry()
+			if !onBlock(blk) {
+				stop = true
+				break
+			}
+		}
+
+		if stop && nextStats {
+			stop = false
+		}
+	}
+}
+
+type StatsBlkIter struct {
+	name       objectio.ObjectName
+	extent     objectio.Extent
+	blkCnt     uint16
+	totalRows  uint32
+	cur        int
+	accRows    uint32
+	curBlkRows uint32
+}
+
+func NewStatsBlkIter(stats *objectio.ObjectStats) *StatsBlkIter {
+	return &StatsBlkIter{
+		name:       stats.ObjectName(),
+		blkCnt:     uint16(stats.BlkCnt()),
+		extent:     stats.Extent(),
+		cur:        -1,
+		accRows:    0,
+		totalRows:  stats.Rows(),
+		curBlkRows: options.DefaultBlockMaxRows,
+	}
+}
+
+func (i *StatsBlkIter) Next() bool {
+	if i.cur >= 0 {
+		i.accRows += i.curBlkRows
+	}
+	i.cur++
+	return i.cur < int(i.blkCnt)
+}
+
+func (i *StatsBlkIter) Entry() *catalog.BlockInfo {
+	if i.cur == -1 {
+		i.cur = 0
+	}
+
+	// assume that all blks have DefaultBlockMaxRows, except the last one
+	if i.cur == int(i.blkCnt-1) {
+		i.curBlkRows = i.totalRows - i.accRows
+	}
+
+	loc := objectio.BuildLocation(i.name, i.extent, i.curBlkRows, uint16(i.cur))
+	blk := &catalog.BlockInfo{
+		BlockID:   *objectio.BuildObjectBlockid(i.name, uint16(i.cur)),
+		SegmentID: i.name.SegmentId(),
+		MetaLoc:   catalog.ObjectLocation(loc),
+	}
+	return blk
+}
+
+func (i *StatsBlkIter) Sequence() uint16 {
+	if i.cur == -1 {
+		i.cur = 0
+	}
+	return uint16(i.cur)
 }
